requirements.txt
google-cloud-aiplatform>=1.25.0
google-auth>=2.6.0 